added regex pre check before building image
