+++ .github/workflows/pr-e2e.yml
@@ -93,6 +93,17 @@ jobs:
           gh pr checkout ${{ needs.triage.outputs.pr_num }}
           git checkout ${{ needs.triage.outputs.commit_sha }}
 
+      - name: Run regex checks
+        env:
+          COMMENT_BODY: ${{ github.event.comment.body }}
+        run: |
+          MESSAGE="$COMMENT_BODY"
+          REGEX='/run-e2e (.+)'
+          if [[ "$MESSAGE" =~ $REGEX ]]
+          then
+            export E2E_TEST_REGEX="$(echo ${BASH_REMATCH[1]} | head -1)"
+          fi
+
       - name: Login to GitHub Container Registry
         uses: docker/login-action@v3
         with:
@@ -158,14 +169,7 @@ jobs:
           ENABLE_OPENTELEMETRY : true
           E2E_IMAGE_TAG: ${{ needs.triage.outputs.image_tag }}
           TEST_CLUSTER_NAME: keda-e2e-cluster-pr
-          COMMENT_BODY: ${{ github.event.comment.body }}
         run: |
-          MESSAGE="$COMMENT_BODY"
-          REGEX='/run-e2e (.+)'
-          if [[ "$MESSAGE" =~ $REGEX ]]
-          then
-            export E2E_TEST_REGEX="$(echo ${BASH_REMATCH[1]} | head -1)"
-          fi
           echo "${{ needs.triage.outputs.pr_num }}"
           make e2e-test
