+++ tests/run-all.go
@@ -7,21 +7,19 @@ import (
 	"context"
 	"fmt"
 	"log"
-	"math/rand"
 	"os"
 	"os/exec"
 	"path"
 	"path/filepath"
-	"regexp"
 	"strings"
 	"sync"
-	"time"
 
 	"golang.org/x/sync/semaphore"
 	"k8s.io/client-go/kubernetes"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 
 	"github.com/kedacore/keda/v2/tests/helper"
+	"github.com/kedacore/keda/v2/tests/utils"
 )
 
 var (
@@ -41,10 +39,10 @@ type TestResult struct {
 func main() {
 	ctx := context.Background()
 
-	e2eRegex := os.Getenv("E2E_TEST_REGEX")
-	if e2eRegex == "" {
-		e2eRegex = ".*_test.go"
-	}
+	//
+	// Get test files
+	//
+	regularTestFiles, sequentialTestFiles := utils.RegexTestFiles()
 
 	//
 	// Install KEDA
@@ -54,17 +52,6 @@ func main() {
 	if !installation.Passed {
 		printKedaLogs()
 		uninstallKeda(ctx)
-		os.Exit(1)
-	}
-
-	//
-	// Detect test cases
-	//
-	regularTestFiles := getRegularTestFiles(e2eRegex)
-	sequentialTestFiles := getSequentialTestFiles(e2eRegex)
-	if len(regularTestFiles) == 0 && len(sequentialTestFiles) == 0 {
-		uninstallKeda(ctx)
-		fmt.Printf("No test has been executed, please review your regex: '%s'\n", e2eRegex)
 		os.Exit(1)
 	}
 
@@ -121,63 +108,6 @@ func executeTest(ctx context.Context, file string, timeout string, tries int) Te
 		fmt.Printf("Execution of %s, attempt %q has failed: %s \n", file, numberToWord(i), err)
 	}
 	return result
-}
-
-func getRegularTestFiles(e2eRegex string) []string {
-	// We exclude utils and sequential folders and helper files
-	filter := func(path string, file string) bool {
-		return !strings.HasPrefix(path, "tests") || // we need this condition to skip non e2e test from execution
-			strings.Contains(path, "utils") ||
-			strings.Contains(path, "sequential") ||
-			!strings.HasSuffix(file, "_test.go")
-	}
-	return getTestFiles(e2eRegex, filter)
-}
-
-func getSequentialTestFiles(e2eRegex string) []string {
-	filter := func(path string, file string) bool {
-		return !strings.HasPrefix(path, "tests") || // we need this condition to skip non e2e test from execution
-			!strings.Contains(path, "sequential") ||
-			!strings.HasSuffix(file, "_test.go")
-	}
-	return getTestFiles(e2eRegex, filter)
-}
-
-func getTestFiles(e2eRegex string, filter func(path string, file string) bool) []string {
-	testFiles := []string{}
-	regex, err := regexp.Compile(e2eRegex)
-
-	if err != nil {
-		return testFiles
-	}
-
-	err = filepath.Walk(".",
-		func(path string, info os.FileInfo, err error) error {
-			if err != nil {
-				return err
-			}
-			// We exclude utils and sequential folders and helper files
-			if filter(path, info.Name()) {
-				return nil
-			}
-			if regex.MatchString(path) {
-				testFiles = append(testFiles, path)
-			}
-
-			return nil
-		})
-
-	if err != nil {
-		return []string{}
-	}
-
-	// We randomize the executions
-	rand.New(rand.NewSource(time.Now().UnixNano()))
-	rand.Shuffle(len(testFiles), func(i, j int) {
-		testFiles[i], testFiles[j] = testFiles[j], testFiles[i]
-	})
-
-	return testFiles
 }
 
 func executeRegularTests(ctx context.Context, testCases []string) []TestResult {
